Fix metadata for /play
- Extract client component to a new file.
